requirements1109.txt
actionlib==1.14.0
angles==1.9.13
argcomplete==3.4.0
astroid==1.6.6
atomicwrites==1.1.5
attrs==23.2.0
autopep8==2.3.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==6.1.0
bondpy==1.8.6
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.2
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.17.0
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
comm==0.2.2
contourpy==1.1.1
controller-manager-msgs==0.20.0
cov-core==1.15.0
coverage==4.5.2
cryptography==2.8
cv-bridge==1.16.2
cycler==0.12.1
dbus-python==1.2.16
decorator==5.1.1
defusedxml==0.7.1
-e git+https://github.com/tonyzhaozh/act.git@742c753c0d4a5d87076c8f69e5628c79a8cc5488#egg=detr&subdirectory=detr
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.0
distro==1.4.0
docutils==0.16
dynamic-reconfigure==1.7.3
einops==0.8.0
empy==3.3.2
entrypoints==0.4
fastjsonschema==2.20.0
filelock==3.13.1
flake8==2.3.0
fonttools==4.53.0
fsspec==2024.2.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
h5py==3.11.0
html5lib==1.0.1
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.0
interactive-markers==1.12.0
ipykernel==5.5.6
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isort==5.13.2
jedi==0.15.2
Jinja2==3.0.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==6.1.12
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
laser_geometry==1.6.7
lazy-object-proxy==1.10.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
message-filters==1.16.0
mistune==0.8.4
more-itertools==4.2.0
moveit-commander==1.1.14
moveit-core==1.1.14
moveit-ros-planning-interface==1.1.14
moveit-ros-visualization==1.1.14
mpmath==1.3.0
nbconvert==5.6.1
nbformat==5.10.4
nbimporter==0.3.4
netifaces==0.10.4
networkx==3.0
nose==1.3.7
nose2==0.9.1
notebook==5.7.16
notify2==0.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
olefile==0.46
osrf-pycommon==2.0.2
packaging==24.1
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pep8==1.7.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==5.5.1
ptyprocess==0.7.0
py==1.8.1
pyassimp==5.2.5
pycairo==1.16.2
pycodestyle==2.12.0
pycryptodomex==3.6.1
pydocstyle==6.3.0
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.18.0
PyGObject==3.36.0
pylint==1.9.4
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==3.1.2
PyQt5==5.14.1
pytest==4.6.9
pytest-cov==2.8.1
python-dateutil==2.9.0.post0
python-gnupg==0.4.5
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python-qt-binding==0.4.4
pytoolconfig==1.3.1
PyYAML==6.0.1
pyzmq==26.0.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.35.1
requests==2.32.3
resource_retriever==1.12.7
roman==2.0.0
rope==1.13.0
ros_numpy==0.0.5
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.24.0
rosdep-modules==0.24.0
rosdistro==0.9.1
rosdistro-modules==0.9.1
rosgraph==1.16.0
rosinstall==0.7.8
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.1
rospkg-modules==1.5.1
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.18.1
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.25
scipy==1.3.3
Send2Trash==1.8.3
sensor-msgs==1.13.1
sip==4.19.21
six==1.16.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
snowballstemmer==2.2.0
soupsieve==1.9.5
srdfdom==0.6.4
supervisor==4.2.5
supervisor-twiddler==1.0.0
sympy==1.13.1
terminado==0.18.1
testpath==0.6.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
tomli==2.0.1
topic-tools==1.16.0
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tornado==6.4.1
tqdm==4.67.0
trac_ik_python==1.6.6
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
ujson==1.35
urdfdom-py==0.4.6
urllib3==2.2.3
vcstools==0.1.42
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.11
wrapt==1.16.0
wstool==0.1.18
xacro==1.14.18
yapf==0.40.2
zipp==3.19.2
